Restructure Django project
- Make the project a little less redundant with the directories
- Slight rename in the project since the directory name did not match
