+++ course_gather/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class CourseViewerConfig(AppConfig):
-    name = 'course_viewer'
+    name = 'course_gather'
