+++ course_gather/settings/dev_settings.py
@@ -1,0 +1,3 @@
+from course_gather.settings.base import *
+SECRET_KEY = ''
+DEBUG = True
