+++ course_gather/settings/prod_settings.py
@@ -1,0 +1,3 @@
+from course_gather.settings.base import *
+DEBUG = False
+ALLOWED_HOSTS = ['*']
