+++ manage.py
@@ -5,7 +5,8 @@ import sys
 
 
 def main():
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'settings.prod_settings')
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE',
+                          'course_gather.settings.local_settings')
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
